+++ gfs-downsized/src/gfs_fc_client.py
@@ -1,24 +1,28 @@
 #!/usr/bin/env python
+
 """
 
 draft version, 2025-02-03
 """
+import sys
 import requests
 import json
 import os
 from time import sleep
-from requests import Response
+from requests import Response, HTTPError
 from multiurl import download
 from datetime import datetime, timedelta, timezone
+from bs4 import BeautifulSoup
 # internal
-from gfs_fc_aux import DATA_DIR, LOG_DIR
+from gfs_fc_aux import DATA_DIR, LOG_DIR, STEPS
 
-COMMON = "{_url}/{_model}.{_yyyymmdd}/{_H}/atmos/{_model}.t{_H}z."
-
-SEMI_LAGRANGIAN_GRID = COMMON + "sfluxgrbf{_fc_hour}.{_extension}"
-GLOBAL_LONGITUDE_LATITUDE_GRID = COMMON + "{_params}{_set}.{_resol}.f{_fc_hour}"
-
-PATTERNS = {
+FC_TIMES = [0, 6, 12, 18]
+COMMON = "{_url}/{_model}.{_yyyymmdd}/{_H}/atmos/"
+SEMI_LAGRANGIAN_GRID = (COMMON
+                        + "{_model}.t{_H}z.sfluxgrbf{_fc_hour}.{_extension}")
+GLOBAL_LONGITUDE_LATITUDE_GRID = (COMMON
+                                  + "{_model}.t{_H}z.{_params}{_set}.{_resol}.f{_fc_hour}")
+PATTERN = {
     "SLS": SEMI_LAGRANGIAN_GRID,
     "GLOB": GLOBAL_LONGITUDE_LATITUDE_GRID
 }
@@ -31,9 +35,9 @@ class Result:
     def __init__(
             self,
             rc,
-            targets
+            target
     ):
-        self.targets = targets
+        self.target = target
         self.rc = rc
 
 
@@ -41,229 +45,271 @@ class Client(object):
     def __init__(
             self,
             *,
+            grid,
+            parameter=None,
+            validity=None,
             model="gfs",  # gdas, enkfgdas
-            grid="GLOB",
             resol="0p25",  # SLS has a resolution of 360 / 1536 !
-            verify=True
+            paramset="",
+            verify=True,
+            **kwargs  # for date & time
     ):
-        self.model = model
+        self.parameter = parameter if parameter else list()
         self.grid = grid
+        self.model = model
         self.resol = resol
+        self.paramset = paramset
         self.verify = verify
+        self.validity = validity if validity else list()
         self.session = requests.Session()
         self.target = "download.grib2"
+        self.date = None
+        self.time = None
         self.lower_by_fc = False
+
+        # define base time at first and check availability
+        self._dateandtime(**kwargs)
+        # if neither date nor time is provided, we get last available fc
+        # or one before otherwise
+        if kwargs.get('date') is None and kwargs.get('time') is None:
+            self._check_availability()
 
     def retrieve(
             self,
-            step,
             *,
-            target=None,
-            resol=None,
+            step,
             **kwargs
     ):
         """
         download grib2 file...
         :param step:
-        :param target:
-        :param resol:
         :param kwargs:
+            target
         :return:
         """
-        expected_size = list()
-        results = list()
-        files = list()
+        target = kwargs.get('target', self.target)
 
-        urls = self._get_urls(
-            step=step,
-            resol=resol,
-            **kwargs)
-        target = target if target else self.target
+        # get m_url for multi-range download
+        m_url = self._get_m_url(step=step)
 
-        for url in urls:
-            print(url)
-            file = "{}{:03d}.grib2".format(
-                target.split(".grib2")[0],
-                step)
-            files.append("{}/{}".format(DATA_DIR, file))  # list of out files
-            expected_size.append(sum([j[1] for j in url['parts']]))
-            results.append(download(
-                **url,
+        file = "{}{:03d}.grib2".format(
+            target.split(".grib2")[0],
+            step)
+
+        if m_url:
+            expected_size = sum([j[1] for j in m_url['parts']])
+            # download byte multirange
+            results = download(
+                **m_url,
                 target="{}/{}".format(DATA_DIR, file),
                 verify=self.verify,
                 session=self.session
-            ))
-            os.chmod("{}/{}".format(DATA_DIR, file), 0o666)  # under Docker owner is root
+            )
+            # under Docker owner is root
+            os.chmod("{}/{}".format(DATA_DIR, file), 0o666)
+        else:
+            sys.exit(1)
 
         return Result(
             rc=expected_size == results,
-            targets=files
+            target="{}/{}".format(DATA_DIR, file)
         )
+
+    @staticmethod
+    def _get_url_paths(
+            *,
+            url: str,
+            ext: str = ".idx",
+            params=None  # ToDo: obsolete
+    ) -> list:
+        """
+        returns list of all index files (full path) of the most recent fc
+        :param url: url of COMMON
+        :param ext: ".idx"
+        :param params: not used
+        :return:
+        """
+        response = requests.get(url, params=params)
+        if response.ok:
+            response_text = response.text
+            soup = BeautifulSoup(response_text, 'html.parser')
+            # list of all downloadable index files per weather forecast time
+            return [url + node.get('href') for node in soup.find_all('a')
+                    if node.get('href').endswith(ext)]
+        else:
+            response.raise_for_status()
+
+    def _check_availability(self) -> None:
+        """
+        checks if all files for the most recent fc are available. If not, rerun
+        dateandtime -6 hrs earlier
+        :return: None
+        """
+        try:
+            idx_list_available = self._get_url_paths(
+                url=self._get_url()
+            )
+            for step in STEPS:
+                if self._get_url(step=step) + ".idx" not in idx_list_available:
+                    raise LookupError
+        except (HTTPError, LookupError):
+            self.lower_by_fc = True
+            print("Files to be downloaded are not available..."
+                  "trying a forecast 6 hrs. earlier.")
+            self._dateandtime()
 
     def _dateandtime(
             self,
             **kwargs
-    ) -> dict:
+    ) -> None:
         """
         set the time at [0|6|12|18] UTC before now. If lower_by_fc scale down
         by -6 hrs.
-        :param lower_by_fc:
         :param kwargs:
-        :return: kwargs (date and time) modified
+        :return:
         """
-        fc_times = [0, 6, 12, 18]
         now = datetime.now(timezone.utc)
         time = int(now.strftime("%H"))
 
         if (kwargs.get('date')
-                and (False if kwargs.get("time") in fc_times else True)):
+                and (False if kwargs.get("time") in FC_TIMES else True)):
             kwargs['time'] = 18  # last fc time of the day
+        else:
+            kwargs['time'] = kwargs.get('time', time - time % 6)
+        assert kwargs['time'] in FC_TIMES, "Value for time (UTC): [0|6|12|18]"
         kwargs['date'] = kwargs.get('date', now.strftime("%Y%m%d"))
-        kwargs['time'] = kwargs.get('time', time - time % 6)
-
-        assert kwargs['time'] in fc_times, "Value for time (UTC): [0|6|12|18]"
 
         if self.lower_by_fc:
             dt = datetime.strptime(
                 f"{kwargs['date']}{kwargs['time']}",
                 "%Y%m%d%H"
             ) - timedelta(hours=6)  # one fc earlier
-            kwargs['date'] = dt.strftime("%Y%m%d")
-            kwargs['time'] = int(dt.strftime("%H"))
+            self.date = dt.strftime("%Y%m%d")
+            self.time = int(dt.strftime("%H"))
+        else:
+            self.date = kwargs['date']
+            self.time = kwargs['time']
 
-        return kwargs
-
-    def _get_urls(
+    def _get_url(
             self,
-            step: int,
-            resol: str,
-            **kwargs
-    ) -> list:
+            step: int = None
+    ) -> str:
         """
         prepare data and configure url string
-        :param kwargs:
+        :param step:
         :return:
         """
-        kwargs = self._dateandtime(**kwargs)
+        args = dict()
 
-        def _configure():
-            args = dict()
-            pattern = PATTERNS[self.grid]
-            args['_url'] = URLS['gfs']
-            args['_model'] = self.model
-            args['_extension'] = "grib2"
-            args['_params'] = "pgrb2"
-            args['_set'] = kwargs.get('paramset', "")
-            args['_resol'] = resol if resol else self.resol
-            args['_yyyymmdd'] = kwargs['date']
-            args['_H'] = "{:02d}".format(kwargs['time'])
-
-            # in case we create multiple urls
-            urls = list()
+        args['_url'] = URLS['gfs']
+        args['_model'] = self.model
+        args['_extension'] = "grib2"
+        args['_params'] = "pgrb2"
+        args['_set'] = self.paramset
+        args['_resol'] = self.resol
+        args['_yyyymmdd'] = self.date
+        args['_H'] = "{:02d}".format(self.time)
+        if step is None:
+            return COMMON.format(**args)
+        else:
             args['_fc_hour'] = "{:03d}".format(step)
-            url = pattern.format(**args)
-            urls.append(url)
+            return PATTERN[self.grid].format(**args)
 
-            return urls
-
-        # end module
-
+    def _get_m_url(
+            self,
+            step: int
+    ) -> dict:
+        """
+        prepare data and configure url string
+        :param step:
+        :return:
+        """
         try:
-            idx = self._call_index(_configure())
-        except Exception as e:  # ToDo specify exception
+            idx = self._call_index(
+                url=self._get_url(step=step)
+            )
+        except (Exception,) as e:
             print(e)
-            print("Resources not available. Revise your parameter set ...")
-            return []
-            # try again for most recent available fc
-            # self.lower_by_fc = True
-            # kwargs = self._dateandtime(**kwargs)
-            # try:
-            #     idx = self._call_index(_configure())
-            # except Exception as e:
-            #     print(e)
-            #     print("Resources not available. Revise your parameter set ...")
-            #     exit(1)
-
-        return self._prepare_request(idx, **kwargs)
+            print("Resource not available. Revise your parameter set ...")
+            return {}
+        return self._prepare_request(idx)
 
     def _call_index(
             self,
-            urls: list[str]
+            url: str
     ) -> dict[str, dict[int, dict]]:
         """
         extract the index files for offset and length of each parameter layer,
         can be filtered by shortName, level, and type.
-        :param urls:
+        :param url:
         :return: index file in dict format
         """
         dix = dict()
         dict_keys = \
             ["offset", "datetime", "shortName", "level", "validity"]
 
-        for url in urls:
-            try:
-                # size of grib data file
-                resp: Response = requests.get(url, stream=True)
-                resp.raise_for_status()
-                length = int(resp.headers.get("Content-length"))
+        try:
+            # total size of grib data file in bytes
+            resp: Response = requests.get(url, stream=True)
+            resp.raise_for_status()
+            length = int(resp.headers.get("Content-length"))
 
-                # and its appropriate index file
-                url_index = "{}.idx".format(url)
-                response = self.session.get(url_index)
-                response.raise_for_status()
-                dix[url] = dict()
-                print(f"Index file {url_index} downloaded")
-            except requests.exceptions.HTTPError as e:
-                raise e  # return empty dict for current url
+            # download its appropriate index file
+            url_index = f"{url}.idx"
+            response = self.session.get(url_index)
+            response.raise_for_status()
+            dix[url] = dict()
+            print(f"Index file {url_index} downloaded")
+        except requests.exceptions.HTTPError as e:
+            raise e  # return empty dict for current url
 
-            for line in response.iter_lines():
-                item = line.decode('utf-8').split(":")[:-1]
-                no = int(item[0])
+        for line in response.iter_lines():
+            item = line.decode('utf-8').split(":")[:-1]
+            no = int(item[0])
 
-                # convert list to dict by merging lists, skip first element that
-                # is number of parameter entry
-                dix[url][no] = dict(zip(dict_keys, item[1:]))
+            # convert list to dict by merging lists, skip first element that
+            # is number of parameter entry
+            dix[url][no] = dict(zip(dict_keys, item[1:]))
 
-                dix[url][no]['datetime'] = \
-                    dix[url][no]['datetime'].replace("d=", "")
-                dix[url][no]['offset'] = int(dix[url][no]['offset'])
-                # replace length by offset minus offset of last record
-                if no > 1:
-                    dix[url][no - 1]['length'] = \
-                        dix[url][no]['offset'] - dix[url][no - 1]['offset']
-                    dix[url][no]['length'] = length - dix[url][no]['offset']
-                # print(json.dumps( dix, indent=2))
+            dix[url][no]['datetime'] = \
+                dix[url][no]['datetime'].replace("d=", "")
+            dix[url][no]['offset'] = int(dix[url][no]['offset'])
+            # replace length by offset minus offset of last record
+            if no > 1:
+                dix[url][no - 1]['length'] = \
+                    dix[url][no]['offset'] - dix[url][no - 1]['offset']
+                dix[url][no]['length'] = length - dix[url][no]['offset']
+            # print(json.dumps( dix, indent=2))
 
-            # caveat: rate limit of <120/minute to the NOMADS site.
+            # Caveat:
+            # NOMAD permits a rate limit of <120/minute to their site.
             # hits are considered to be head/listing commands as well as actual
-            # data download attempts, i.e. each get request
+            # data download attempts, i.e. each get http request
             # The block is temporary and typically lasts for 10 minutes
             # the system is automatically configured to blacklist an IP if it
             # continually hits the site over the threshold.
             # source: ncep.pmb.dataflow@noaa.gov (Brian)
-            sleep(.5)
-        # end for loop url
+            # Hence, configure the sleep argument accordingly!
+        sleep(1.)
 
         with open("{}/indices.json".format(LOG_DIR), "w") as log_handle:
             json.dump(dix, log_handle, indent=2)
-        os.chmod("{}/indices.json".format(LOG_DIR), 0o666)  # docker owner is root, anyone can delete
+        # docker owner is root, anyone can delete
+        os.chmod("{}/indices.json".format(LOG_DIR), 0o666)
 
         return dix
 
-    @staticmethod
     def _prepare_request(
-            idx: dict[str, dict[int, dict]],
-            **kwargs
-    ) -> list:
+            self,
+            idx: dict[str, dict[int, dict]]
+    ) -> dict:
         """
         prepare data multirange downloads of single url, see
         https://github.com/ecmwf/multiurl
         :param idx:
-        :param kwargs:
         :return:
         """
-        url_download = list()
-        parameter: list = kwargs.get('parameter')
+        url_download = dict()
 
         for url, v in idx.items():
             t = tuple()
@@ -271,16 +317,17 @@ class Client(object):
             highest_id = sorted(v, key=lambda x: x, reverse=True)[0]
             size = v[highest_id]['offset'] + v[highest_id]['length']
 
-            if not parameter:  # download entire parameter set
-                url_download.append(
-                    {"url": url,
-                     "parts": ((0, size),)}
-                )
+            # download entire parameter set, not recommended
+            if not self.parameter:
+                url_download = {
+                    "url": url,
+                    "parts": ((0, size),)
+                }
                 continue
 
-            # number of parameter is key
+            # number of parameter is key, might considere number as viable key!
             for k, value in v.items():
-                for p in parameter:  # parameter to be selected
+                for p in self.parameter:  # parameter to be selected
                     predicate = False
                     # checks
                     assert p.get('shortName'), "shortName must not be empty!"
@@ -300,9 +347,9 @@ class Client(object):
                         else:
                             predicate = True
 
-                    if predicate and kwargs.get('validity'):
+                    if predicate and self.validity:
                         if not any(i in value['validity']
-                                   for i in kwargs['validity']):
+                                   for i in self.validity):
                             predicate = False
                     if predicate:
                         print("{}:{}:{}:{}".format(
@@ -318,10 +365,9 @@ class Client(object):
                 # remove duplicates and sort according to offset, but
                 # slow with big numbers
                 t = tuple(sorted(set(t), key=t.index))
-                url_download.append(
-                    {"url": url,
-                     "parts": t}
-                )
+                url_download = {
+                    "url": url,
+                    "parts": t}
             else:
                 raise KeyError("No filter applied.")
             # end for loop item number each url